+++ src/php/Board/Exception/UnknownBoardTypeException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace FreeElephants\HexoNards\Board\Exception;
+
+use FreeElephants\HexoNards\Exception\OutOfBoundsException;
+
+/**
+ * @author samizdam <samizdam@inbox.ru>
+ */
+class UnknownBoardTypeException extends OutOfBoundsException
+{
+
+}