+++ src/php/Exception/OutOfBoundsException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace FreeElephants\HexoNards\Exception;
+
+/**
+ * @author samizdam <samizdam@inbox.ru>
+ */
+class OutOfBoundsException extends \OutOfBoundsException implements ExceptionInterface
+{
+
+}